requirements.txt
Django
yagmail
python-telegram-bot
requests
django-heroku
gunicorn

